packages/ml-services/requirements.txt
# ML Service Dependencies
ultralytics>=8.0.0          # YOLOv8
opencv-python>=4.8.0        # OpenCV
fastapi>=0.104.0            # REST API framework
uvicorn[standard]>=0.24.0   # ASGI server
python-dotenv>=1.0.0        # Environment variable management
numpy>=1.24.0               # Numerical computing
